third_party/blink/web_tests/virtual/blur-on-remove/fast/dom/shadow/shadow-removechild-and-blur-event-expected.txt

Check focused shadow node dispatchs blur event on removeChild of host.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS logs.toString() is "start-removeChild,blur,end-removeChild"
PASS successfullyParsed is true

TEST COMPLETE
